+++ models/models.go
@@ -1,0 +1,11 @@
+package models
+
+type ResolveBody struct {
+	ID            int    `json:"id"`
+	LocationType  int    `json:"type"`
+	NewAddress    string `json:"new_address"`
+	OpenAddress   string `json:"open_address"`
+	Apartment     string `json:"apartment"`
+	Reason        string `json:"reason"`
+	TweetContents string `json:"tweet_contents"`
+}
